requirements.txt
urllib3==1.22
firebase_admin==2.8.0
python-dotenv==0.7.1
Flask==0.12.2
gunicorn==19.7.1